+++ wdm-ai-botkit-extension-license/wdm-ai-botkit-extension.php
@@ -9,7 +9,7 @@
  * that starts the plugin.
  *
  * @link              https://wisdmlabs.com
- * @since             1.0.0
+ * @since             1.0.1
  * @package           Wdm_Ai_Botkit_Extension
  *
  * @wordpress-plugin